add connection notifications on submit error
This change adds some logic in the front-end to give a user
notifications when there is trouble getting prediction information from
the server. If the server returns an empty response or if the connection
fails, a notification will be displayed to the user.

Also adds a convenience method to clear the table when pressing submit.